use RGBA8 when extracting original images

+++ jsettlers.graphics/src/main/java/jsettlers/graphics/image/reader/ImageArrayProvider.java
@@ -40,5 +40,5 @@ public interface ImageArrayProvider {
 	 */
 	void startImage(int width, int height) throws IOException;
 
-	void writeLine(short[] data, int length) throws IOException;
+	void writeLine(int[] data, int length) throws IOException;
 }
